update to v0.3.0
added justify whitespace, allowing characters to stay together, spreading only the whitespace.

fixed the justify characters to include whitespace characters as well to spread more evenly.

reverted renaming - usage of Align reverted back to Alignment

each line now also tracks the index of the first character in the string that begins the line as well as the length of that line in characters (excluding the newline character). this was necessary to include spaces in justification.

separated the justification quad modifying codes from the basic alignment quad modifying code.