update to v0.4
added ability to set line height multiplier (line height becomes multiplied by the given value so 1 is normal height, 0 is no height and 2 is double height - an empty line inbetween)

added ability to set a letter spacing. this value is a multiple of the width of a space. letter spacing is most commonly specified as an addition space so this value is 0 as default - no extra spacing. a value of 1 would add the width of a space to each character whereas a value of -1 would subtract the width of a space to each character. note that SFML Text's letter spacing is a quite strange value: it is positive (1) when no extra spacing is applied and a third of a width of a space per each whole number from that. this means that 0 would be 1/3 of a width of a space from each character's normal spacing and 2 would be 1/3 of a width of a space as letter space (space added in addition to normal character placement). when creating an Aline object from (or setting it equal to) an SFML Text object, the letter spacing value is automatically corrected and they will still match.

added local and global bounds.

added ability to round (x) position values when glyphs are moved due to alignment. they are rounded to the integer to keep rendering of them away from fractional values. customisable offsets can still be fractional, however, and these are not rounded, giving you flexibility to offset them by a fractional amount but still have them rounded first. rounding can also be disabled but it's enabled by default.

removed some unused lines.